+++ app/api/dto/event_dto.go
@@ -7,9 +7,9 @@ import (
 )
 
 type EntityEventDTO struct {
-	ID         int64     `json:"id"`
-	IdentityId *int64    `json:"identity_id,omitempty"`
-	EntityId   *int64    `json:"entity_id,omitempty"`
+	ID         int       `json:"id"`
+	IdentityId *int      `json:"identity_id,omitempty"`
+	EntityId   *int      `json:"entity_id,omitempty"`
 	EntityType *string   `json:"entity_type,omitempty"`
 	EventType  string    `json:"event_type"`
 	Data       string    `json:"data"`
