+++ app/services/entity_event_service.go
@@ -10,8 +10,8 @@ import (
 
 type EntityEventService interface {
 	Filter(ctx context.Context, query any, args ...any) ([]models.EntityEvent, error)
-	ListByEntity(ctx context.Context, entityId int64, entityType string) ([]models.EntityEvent, error)
-	ListByEntityIds(ctx context.Context, entityIds []int64, entityType string) ([]models.EntityEvent, error)
+	ListByEntity(ctx context.Context, entityId int, entityType string) ([]models.EntityEvent, error)
+	ListByEntityIds(ctx context.Context, entityIds []int, entityType string) ([]models.EntityEvent, error)
 }
 
 type entityEventService struct {
@@ -38,12 +38,12 @@ func (s *entityEventService) Filter(ctx context.Context, query any, args ...any)
 }
 
 // ListByEntity implements [EntityEventService].
-func (s *entityEventService) ListByEntity(ctx context.Context, entityId int64, entityType string) ([]models.EntityEvent, error) {
+func (s *entityEventService) ListByEntity(ctx context.Context, entityId int, entityType string) ([]models.EntityEvent, error) {
 	return s.repository.ListByEntity(ctx, entityId, entityType)
 }
 
 // ListByEntityIds implements [EntityEventService].
-func (s *entityEventService) ListByEntityIds(ctx context.Context, entityIds []int64, entityType string) ([]models.EntityEvent, error) {
+func (s *entityEventService) ListByEntityIds(ctx context.Context, entityIds []int, entityType string) ([]models.EntityEvent, error) {
 	if len(entityIds) == 0 {
 		return []models.EntityEvent{}, nil
 	}
